backend/requirements.txt
requests==2.26.0
Django==3.2
djangorestframework==3.12.4
djangorestframework-simplejwt==5.2.2
djoser==2.2.0
PyJWT==2.1.0
django-filter==22.1
Pillow==9.5.0
python-dotenv==0.19.2